+++ packages/react/src/examples/forms/validationHybrid.stories.tsx
@@ -9,7 +9,7 @@ import {
   Button,
   Checkbox,
   DateInput,
-  ErrorSummary,
+  Notification,
   PhoneInput,
   RadioButton,
   Select,
@@ -114,7 +114,7 @@ export const Hybrid = () => {
    */
   const renderErrorSummary = () =>
     hasBackendErrors ? (
-      <ErrorSummary label="Form contains following errors" className="hds-example-form__error-summary" autofocus>
+      <Notification type="error" label="Form contains following errors" style={{ marginBottom: 'var(--spacing-m)' }}>
         <ul>
           <li>
             Error 1:{' '}
@@ -123,7 +123,7 @@ export const Hybrid = () => {
             </a>
           </li>
         </ul>
-      </ErrorSummary>
+      </Notification>
     ) : null;
 
   return (