+++ packages/react/src/examples/forms/validationStatic.stories.tsx
@@ -9,7 +9,7 @@ import {
   Button,
   Checkbox,
   DateInput,
-  ErrorSummary,
+  Notification,
   PhoneInput,
   RadioButton,
   SelectionGroup,
@@ -83,7 +83,7 @@ export const Static = () => {
    */
   const renderErrorSummary = () =>
     Object.keys(errors).length > 0 ? (
-      <ErrorSummary label="Form contains following errors" className="hds-example-form__error-summary" autofocus>
+      <Notification type="error" label="Form contains following errors" style={{ marginBottom: 'var(--spacing-m)' }}>
         <ul>
           {Object.keys(errors).map((errorKey, index) => (
             <li key={`error-${errorKey}`}>
@@ -92,7 +92,7 @@ export const Static = () => {
             </li>
           ))}
         </ul>
-      </ErrorSummary>
+      </Notification>
     ) : null;
 
   return (